PyTorch Operator: Move manifests development upstream (#320)
As part of the work of wg-manifests for 1.3
(https://github.com/kubeflow/manifests/issues/1735), we are moving manifests
development in upstream repos. This gives the application developers full
ownership of their manifests, tracked in a single place.

This commit copies the manifests for application `PyTorch Operator`
from path `apps/pytorch-job/upstream` of kubeflow/manifests to path
`manifests` of the upstream repo (https://github.com/kubeflow/pytorch-operator).

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>